[Random Node] Problem 4.11 of Cracking the Coding Interview (6th Edition)
The problem was solved very quickly. However, this exposed a limit in my
understanding regarding Binary Trees and clojure.zip. I realized that with my
data structures (trees) of vectors of the form [<value> <left-node>
<right-node>] I was causing zip/next and zip/down to also run across the nil
nodes (if they exist). I did not notice this on the older problems.

If you want to randomly select a node multiple times from a tree, then you
should add a NODE_CHILDREN_COUNT property to each node. This is the number of
nodes (not including yourself, so at minimum 1) that are within this tree. Then
you can use rand-int and the summation of all children to recursively grab a
random node.

Note: I have some concerns that this may not be perfectly uniform.